complete rewrite, but still issues
Main issue: switching from playing an audiobook to a music item kills 
the player detection

But you can refresh the page anytime you start playing something new.  

Or you can switch between items of the same type (from audiobook to 
audiobook / music to music) (at least a couple times, then hoopla seems 
to have a separate issue that chokes)

For me, this works for now - but it's not good.  

Also this re-write fixes a discrepancy between how web-scrobbler parses 
ArtistTrack in regards to hoopla which doesn't actually have ArtistTrack 
text, but AlbumArtist text instead.  (and sometimes the Albums have 
hyphens in them which throws off the Utils detection) So, this avoids 
the Utils splitter and prescribes a custom method for hoopla